refactor: simplify error pages
