+++ webapp/client/apps/orchestration-cluster-webapp/src/tasklist/pages/TaskDetailsProcessError.test.tsx
@@ -10,45 +10,38 @@ import {render} from 'vitest-browser-react';
 import {describe, expect, vi} from 'vitest';
 import {userEvent} from 'vitest/browser';
 import {it} from '#/vitest-modules/test-extend';
-import {TaskDetailsProcessError} from './TaskDetailsProcessError';
+import {TaskDetailsProcessRouteError} from './TaskDetailsProcessError';
+
+function failedResponseError(status: number) {
+	return {variant: 'failed-response', response: new Response(null, {status}), networkError: null} as unknown as Error;
+}
 
 describe('<TaskDetailsProcessError />', () => {
 	it('should explain when the user does not have permission to view the process', async () => {
-		const screen = await render(<TaskDetailsProcessError variant="forbidden" />);
+		const screen = await render(
+			<TaskDetailsProcessRouteError error={failedResponseError(403)} info={{componentStack: ''}} reset={vi.fn()} />,
+		);
 
 		await expect.element(screen.getByText("You don't have permission to view the process")).toBeVisible();
 		await expect
 			.element(screen.getByText("You don't have the necessary permissions. Contact your admin to request access."))
 			.toBeVisible();
-	});
-
-	it('should not offer retry when access is forbidden', async () => {
-		const screen = await render(<TaskDetailsProcessError variant="forbidden" onRetry={vi.fn()} />);
-
 		await expect.element(screen.getByRole('button', {name: 'Try again'})).not.toBeInTheDocument();
 	});
 
-	it('should explain when the process diagram cannot be loaded', async () => {
-		const screen = await render(<TaskDetailsProcessError variant="generic" />);
+	it('should allow users to retry when the process diagram cannot be loaded', async () => {
+		const reset = vi.fn();
+		const screen = await render(
+			<TaskDetailsProcessRouteError error={failedResponseError(500)} info={{componentStack: ''}} reset={reset} />,
+		);
 
 		await expect.element(screen.getByText('Process could not be loaded')).toBeVisible();
 		await expect
 			.element(screen.getByText('The BPMN process diagram could not be loaded. Please try again later.'))
 			.toBeVisible();
-	});
-
-	it('should allow users to retry loading the process diagram when retry is available', async () => {
-		const onRetry = vi.fn();
-		const screen = await render(<TaskDetailsProcessError variant="generic" onRetry={onRetry} />);
 
 		await userEvent.click(screen.getByRole('button', {name: 'Try again'}));
 
-		expect(onRetry).toHaveBeenCalledOnce();
-	});
-
-	it('should not offer retry when retry is unavailable', async () => {
-		const screen = await render(<TaskDetailsProcessError variant="generic" />);
-
-		await expect.element(screen.getByRole('button', {name: 'Try again'})).not.toBeInTheDocument();
+		expect(reset).toHaveBeenCalledOnce();
 	});
 });
